fix(ipc): foundation review remediation
Addresses the full ipc-codegen/ipc-runtime review (68 findings):
- runtime: NAPI lifecycle (close/join, TSFN leak, sync deadline), zero-length
  response deadlock, u64 timeout truncation, max-frame guards, crash cleanup,
  unified constants, UDS timeouts, new cpp/rust/ts test suites
- codegen: schema validation (required error variant, name pairing, reserved
  words), unified server error wrapping, zig comptime handler injection,
  uniform --strip-method-prefix, skeleton/dead-code removal
- wire fixes: rust Option<bytes>/[bytes;N] serde, ts u64 bigint bridging +
  bin32 guards, zig signed-int tolerance, cpp union ordering + PrimAlias so
  schema reflection of generated types round-trips aliases
- tests: golden corpus extended (error variant pinned) with cpp/zig runners,
  error-path + boundary-value steps in every matrix combo, FFI compile checks

+++ ipc-runtime/cpp/ipc_runtime/shm/mpsc_shm.hpp
@@ -70,7 +70,7 @@ class MpscConsumer {
      * @param timeout_ns Total timeout in nanoseconds (spins 10ms, then futex waits for remainder)
      * @return Ring index with data, or -1 on timeout
      */
-    int wait_for_data(uint32_t timeout_ns);
+    int wait_for_data(uint64_t timeout_ns);
 
     /**
      * @brief Peek data from specific ring
@@ -79,7 +79,7 @@ class MpscConsumer {
      * @param timeout_ns Timeout in nanoseconds
      * @return Pointer to data, or nullptr on timeout
      */
-    void* peek(size_t ring_idx, size_t want, uint32_t timeout_ns);
+    void* peek(size_t ring_idx, size_t want, uint64_t timeout_ns);
 
     /**
      * @brief Release data from specific ring
@@ -134,7 +134,7 @@ class MpscProducer {
      * @param timeout_ns Timeout in nanoseconds
      * @return Pointer to buffer, or nullptr on timeout
      */
-    void* claim(size_t want, uint32_t timeout_ns);
+    void* claim(size_t want, uint64_t timeout_ns);
 
     /**
      * @brief Publish data to producer's ring (rings doorbell)
